exception 처리 구현 수정

+++ src/main/java/com/sparta/hanghaeboardproject2/service/BoardService.java
@@ -3,6 +3,7 @@ package com.sparta.hanghaeboardproject2.service;
 import com.sparta.hanghaeboardproject2.dto.BoardDto;
 import com.sparta.hanghaeboardproject2.model.Board;
 import com.sparta.hanghaeboardproject2.repository.BoardRepository;
+import com.sparta.hanghaeboardproject2.security.MemberDetailsImpl;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
@@ -42,11 +43,20 @@ public class BoardService {
         Board foundBoard = boardRepository.findById(id).orElseThrow(
                 () -> new IllegalArgumentException("아이디가 존재하지 않습니다.")
         );
+
         foundBoard.update(boardDto);
         return foundBoard.getId();
     }
 
-    public void deleteBoard(Long id) {
+    public void deleteBoard(Long id, MemberDetailsImpl memberDetails) {
+        Board board = boardRepository.findById(id).orElseThrow(
+                () -> new IllegalArgumentException("해당 id를 찾을 수 없습니다.")
+        );
+
+        if (!board.getWriter().equals(memberDetails.getUsername())) {
+            throw new IllegalArgumentException("해당 게시글 작성자만 삭제할 수 있습니다.");
+        }
+
         boardRepository.deleteById(id);
     }
 }
